more things

+++ app/Http/Controllers/UserController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\UserUpdateRequest;
 use App\Http\Resources\UserResource;
 use App\Models\User;
 use Auth;
+use Gate;
 use Hash;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -21,6 +22,7 @@ class UserController extends Controller
      */
     public function index()
     {
+        Gate::authorize('view', 'users');
        $user = User::paginate();
 
        return response(UserResource::collection($user), Response::HTTP_ACCEPTED);
@@ -34,6 +36,7 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('edit', 'users');
        $user = User::create( $request->only('first_name', 'last_name', 'email', 'role_id') +
           [
             'password' => Hash::make(1234)
@@ -50,6 +53,7 @@ class UserController extends Controller
      */
     public function show($id)
     {
+        Gate::authorize('view', 'users');
        $user = User::find($id);
 
         return response(new UserResource($user), Response::HTTP_ACCEPTED);
@@ -64,6 +68,7 @@ class UserController extends Controller
      */
     public function update(UserUpdateRequest $request, $id)
     {
+        Gate::authorize('edit', 'users');
         $user = User::find($id);
 
         $user->update($request->only('first_name', 'last_name', 'email', 'role_id')
@@ -80,6 +85,7 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('view', 'users'); 
         $user = User::destroy($id);
 
         return response(null, Response::HTTP_NO_CONTENT);
@@ -89,7 +95,9 @@ class UserController extends Controller
        
         $user = Auth::user();
 
-        return response(new UserResource($user), Response::HTTP_ACCEPTED);
+        return (new UserResource($user))->additional([
+            'permissions' => $user->permissions()
+        ]);
     }
 
     public function updateInfo(UpdateUserInfoRequest $request){
